requirements.txt
flask
flask-cors
tensorflow
keras-tuner
python-dotenv
joblib
requests
seaborn
pandas-ta
keras
numpy
pandas
scikit-learn
matplotlib
yfinance
lxml
pyarrow